pyproject.toml
[tool.poetry]
name = "cisco-chat-backend"
version = "1.1.0"
description = "This is an AI chatbot that uses OpenAI and Security Advisories data to interact with the natural language."
license = "MIT"
authors = [
    "Candido Gomes <candidosg@gmail.com>",
]
maintainers = [
    "Candido Gomes <candidosg@gmail.com>",
]
readme = "README.md"
homepage = "https://github.com/candidosales/cisco-chat-backend"
repository = "https://github.com/candidosales/cisco-chat-backend"
keywords = ["cisco", "chatgtp", "openai", "langchain"]
packages = [{include = "app"}]

[tool.poetry.scripts]
start = "app.main:start"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
python-dotenv = "1.0.0"
embedchain = "0.1.64"
ruff = "0.1.13"
pandas = "2.1.4"
fastapi = "0.109.0"
pydantic = "2.5.3"
beautifulsoup4 = "4.12.2"
langchain = "0.0.336"
chromadb = "0.4.22"
tiktoken = "0.4.0"
pre-commit = "^3.6.0"
modal-client = "0.56.4584"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"